Update new transaction status view

+++ KyberNetwork/KyberNetwork/Extensions/UIView+Kyber.swift
@@ -131,7 +131,7 @@ extension UIView {
       animate.duration = duration
       animate.repeatCount = Float.infinity
       animate.fromValue = 0.0
-      animate.toValue = Float.pi * 2.0
+      animate.toValue = CGFloat.pi * 2.0
       self.layer.add(animate, forKey: kAnimationKey)
     }
   }
